refactor(cursor-usage): modularize billing export
